Migrate `Components/Icons` story to TS
