+++ packages/livechat/babel.config.js
@@ -4,13 +4,11 @@ module.exports = {
 		['@babel/plugin-transform-class-properties', { loose: true }],
 		['@babel/plugin-transform-private-methods', { loose: true }],
 		['@babel/plugin-transform-private-property-in-object', { loose: true }],
-		['@babel/plugin-transform-react-jsx', { pragma: 'h', pragmaFrag: 'Fragment' }],
 		[
-			'babel-plugin-jsx-pragmatic',
+			'@babel/plugin-transform-react-jsx',
 			{
-				module: 'preact',
-				import: 'h',
-				export: 'h',
+				runtime: 'automatic',
+				importSource: 'preact',
 			},
 		],
 	],
